deployments/operator/otel-tracing/requirements.txt
# Traceloop SDK (OpenLLMetry)

traceloop-sdk
